requirements.txt
asgiref==3.6.0
certifi==2022.12.7
chardet==3.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-appconf==1.0.5
django-cart==1.0.4
django-crispy-forms==1.8.1
django-email-users==1.1.0
django-extensions==3.2.1
django-formtools==2.1
django-imagekit==4.1.0
django-import-export==2.0
django-redis==4.10.0
django-reversion==3.0.5
django-simpleui==2023.3.1
djangorestframework==3.10.2
et-xmlfile==1.1.0
idna==2.8
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.2
pilkit==2.0
Pillow==9.5.0
PyMySQL==1.0.3
pytz==2023.3
PyYAML==6.0
redis==3.3.8
requests==2.22.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
urllib3==1.25.11
xlrd==2.0.1
xlwt==1.3.0
